requirements.txt
matplotlib==3.4.1
numpy==1.20.2
tensorflow==2.8.0
Pillow==9.0.1
opencv-python==4.5.5.62
ray==1.12.0
